Add method in RegistratorService for input/saving resource in database

+++ src/main/java/org/registrator/community/dao/implementation/DaoOperationsImp.java
@@ -66,7 +66,7 @@ public class DaoOperationsImp<T> implements IDaoOperations<T> {
 
 	@SuppressWarnings("unchecked")
 	@Override
-	public T findById(Long entityId) {
+	public T findById(Integer entityId) {
 
 		Session session = null;
 		T element=null;
